refactor(MBService): Decouple the code responsible from generate the MB reference from the Response
